refactor: move to modular types
