+++ .gitignore
@@ -38,3 +38,5 @@ functions/lib/
 *.tsbuildinfo
 *.js.map
 *.d.ts.map
+
+.vscode
