--- functions/src/types/index.ts
@@ -1,8 +1,0 @@
-// Re-export all types from individual files
-export * from "./schemas";
-export * from "./user";
-export * from "./email";
-export * from "./calendar";
-export * from "./resend";
-export * from "./openai";
-export * from "./common";
