+++ functions/src/util/types.ts
@@ -1,4 +1,53 @@
-import {AttachmentInfo, TransformedEmail} from "./email";
+import {Request} from "express";
+import {AttachmentInfo, TransformedEmail} from "../agents/calendar/types";
+
+// ============================================================================
+// COMMON TYPES
+// ============================================================================
+
+// Error with code
+export interface ErrorWithCode extends Error {
+  code?: number | string;
+}
+
+// Express request with query params
+export interface RequestWithQuery extends Request {
+  query: {
+    uuid?: string;
+    uid?: string;
+    eventId?: string;
+    attendees?: string | string[];
+    calendarId?: string;
+    code?: string;
+  };
+}
+
+// Task dispatch options
+export interface TaskDispatchOptions {
+  functionName: string;
+  data: unknown;
+  deadline?: number;
+  scheduleDelaySeconds?: number;
+  location?: string;
+}
+
+// Async result tuple
+export type AsyncResult<T> = [Error | null, T | null];
+
+// Analytics event params
+export interface AnalyticsEventParams {
+  traffic_type?: string;
+  action?: string;
+  reason?: string;
+  error_type?: string;
+  result?: string;
+  operation?: string;
+  [key: string]: string | undefined;
+}
+
+// ============================================================================
+// RESEND API TYPES
+// ============================================================================
 
 // Resend email options
 export interface ResendEmailOptions {
@@ -101,4 +150,23 @@ export interface MockSentEmail {
   text?: string;
   html: string;
   headers: Record<string, string>;
+}
+
+// Test data for mock Resend client
+export interface TestData {
+  emailContent: TransformedEmail | ResendEmailData;
+  attachmentsList: AttachmentInfo[];
+}
+
+// Task queue request wrapper
+export interface TaskRequest {
+  data: ResendWebhookData;
+}
+
+// Task dispatch result
+export interface DispatchResult {
+  message: string;
+  data?: unknown;
+  sentEmail?: unknown;
+  error?: string;
 }
